requirements.txt
zulip==0.8.2
openai==0.27.2
python-dotenv==1.0.0
